requirements.txt
absl-py==2.2.2
astunparse==1.6.3
cachetools==5.5.2
certifi==2023.11.17
charset-normalizer==3.3.2
einops==0.7.0
filelock==3.13.1
fsspec==2023.12.2
google-auth==2.40.1
google-auth-oauthlib==1.0.0
grpcio==1.70.0
huggingface-hub==0.20.1
idna==3.6
imageio==2.35.1
importlib_metadata==8.5.0
jittor==1.3.10.0
joblib==1.4.2
lazy_loader==0.4
lmdb==1.6.2
Markdown==3.7
MarkupSafe==2.1.5
natsort==8.4.0
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
ruff==0.12.8
safetensors==0.4.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tifffile==2023.7.10
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.6
zipp==3.20.2
